created alternative branch

+++ src/components/ItemList.js
@@ -9,9 +9,24 @@ class ItemList extends Component {
         <div style={{ height: '100%', display: 'flex', flexDirection: 'column', alignItems: 'center', justifyContent: 'space-around' }}>
           <div style={{ width: '100%', display: 'flex', flexDirection: 'column', alignItems: 'center' }}>
             <h3>OSTOSLISTA</h3>
-            {this.props.stateItemsForItemList.map(item => (
-              <Item key={item.id} itemForItem={item} markComplete={this.props.markComplete} deleteItem={this.props.deleteItem} />
-            )
+            {this.props.stateItemsForItemList.map(item => {
+              let moi;
+              if (!item.collected) {
+                moi = <Item key={item.id} itemForItem={item} markComplete={this.props.markComplete} deleteItem={this.props.deleteItem}/>
+              }
+              return moi;
+            }
+            )}
+          </div>
+          <div style={{ width: '100%', display: 'flex', flexDirection: 'column', alignItems: 'center' }}>
+            <h3>KERÄTYT</h3>
+            {this.props.stateItemsForItemList.map(item => {
+              let tere;
+              if (item.collected) {
+                tere = <Item key={item.id} itemForItem={item} markComplete={this.props.markComplete} deleteItem={this.props.deleteItem}/>
+              }
+              return tere;
+            }
             )}
           </div>
         </div>
